+++ main.go
@@ -5,8 +5,7 @@ import (
 	"net/http"
 
 	"github.com/diegosorrilha/users-api/configs"
-	"github.com/diegosorrilha/users-api/handlers"
-	"github.com/go-chi/chi/v5"
+	"github.com/diegosorrilha/users-api/routers"
 )
 
 func main() {
@@ -17,14 +16,7 @@ func main() {
 
 	server_path := fmt.Sprintf("localhost:%v", configs.GetServerPort())
 
-	r := chi.NewRouter()
-
-	// routes
-	r.Get("/users", handlers.ListUsers)
-	r.Post("/users", handlers.CreateUser)
-	r.Get("/users/{id}", handlers.GetUser)
-	r.Delete("/users/{id}", handlers.DeleteUser)
-	r.Put("/users/{id}", handlers.UpdateUser)
+	r := routers.CreateNewRouter()
 
 	fmt.Printf("Server running: http://%v/users", server_path)
 	http.ListenAndServe(server_path, r)